[tracker] refactoring
