+++ hal/inc/hal_platform.h
@@ -37,7 +37,11 @@
 #define HAL_PLATFORM_WIFI 0
 #endif /* HAL_PLATFORM_WIFI */
 
-#if HAL_PLATFORM_WIFI && HAL_PLATFORM_NCP_AT
+#ifndef HAL_PLATFORM_WIFI_NCP_SDIO
+#define HAL_PLATFORM_WIFI_NCP_SDIO (0)
+#endif // HAL_PLATFORM_WIFI_NCP_SDIO
+
+#if HAL_PLATFORM_WIFI && HAL_PLATFORM_NCP_AT && !HAL_PLATFORM_WIFI_NCP_SDIO
 # ifndef HAL_PLATFORM_WIFI_SERIAL
 #  error "HAL_PLATFORM_WIFI_SERIAL is not defined"
 # endif /* HAL_PLATFORM_WIFI_SERIAL */
@@ -425,5 +429,13 @@
 #ifndef HAL_PLATFORM_MULTIPART_SYSTEM
 #define HAL_PLATFORM_MULTIPART_SYSTEM (0)
 #endif // HAL_PLATFORM_MULTIPART_SYSTEM
+
+#ifndef HAL_PLATFORM_NCP_COUNT
+#define HAL_PLATFORM_NCP_COUNT (1)
+#endif // HAL_PLATFORM_NCP_COUNT
+
+#ifndef HAL_PLATFORM_WIFI_COMPAT
+#define HAL_PLATFORM_WIFI_COMPAT (0)
+#endif // HAL_PLATFORM_WIFI_COMPAT
 
 #endif /* HAL_PLATFORM_H */
