+++ hal/inc/hal_platform_compat.h
@@ -22,6 +22,7 @@
 
 #if PLATFORM_ID <= PLATFORM_P1 || PLATFORM_ID == PLATFORM_NEWHAL
     #define HAL_PLATFORM_WIFI 1
+    #define HAL_PLATFORM_WIFI_COMPAT 1
 #endif
 
 #if PLATFORM_ID == PLATFORM_ELECTRON_PRODUCTION
