+++ hal/network/ncp_client/esp32/esp32_ncp_client.cpp
@@ -101,27 +101,20 @@ int Esp32NcpClient::init(const NcpClientConfig& conf) {
 #endif
     espOff();
 
-    // FIXME: should we add a configuration for the communication interface instead of using conditional compilation
-#if PLATFORM_ID == PLATFORM_ARGON || PLATFORM_ID == PLATFORM_ASOM
+#if !HAL_PLATFORM_WIFI_NCP_SDIO
     // Initialize serial stream
     std::unique_ptr<SerialStream> serial(new(std::nothrow) SerialStream(HAL_USART_SERIAL2, ESP32_NCP_DEFAULT_SERIAL_BAUDRATE,
             SERIAL_8N1 | SERIAL_FLOW_CONTROL_RTS_CTS));
+#else
+    std::unique_ptr<SdioStream> serial(new(std::nothrow) SdioStream(HAL_SPI_INTERFACE2, ESP32_NCP_DEFAULT_SDIO_SPEED, WIFI_CS, WIFI_INT));
+#endif // !HAL_PLATFORM_WIFI_NCP_SDIO
     CHECK_TRUE(serial, SYSTEM_ERROR_NO_MEMORY);
-#elif PLATFORM_ID == PLATFORM_TRACKER
-    std::unique_ptr<SdioStream> sdio(new(std::nothrow) SdioStream(HAL_SPI_INTERFACE2, ESP32_NCP_DEFAULT_SDIO_SPEED, WIFI_CS, WIFI_INT));
-    CHECK_TRUE(sdio, SYSTEM_ERROR_NO_MEMORY);
-#endif
     // Initialize muxed channel stream
     decltype(muxerAtStream_) muxStrm(new(std::nothrow) decltype(muxerAtStream_)::element_type(&muxer_, ESP32_NCP_AT_CHANNEL));
     CHECK_TRUE(muxStrm, SYSTEM_ERROR_NO_MEMORY);
     CHECK(muxStrm->init(ESP32_NCP_AT_CHANNEL_RX_BUFFER_SIZE));
-#if PLATFORM_ID == PLATFORM_ARGON || PLATFORM_ID == PLATFORM_ASOM
     CHECK(initParser(serial.get()));
     serial_ = std::move(serial);
-#elif PLATFORM_ID == PLATFORM_TRACKER
-    CHECK(initParser(sdio.get()));
-    sdio_ = std::move(sdio);
-#endif
     muxerAtStream_ = std::move(muxStrm);
     conf_ = conf;
     ncpState_ = NcpState::OFF;
@@ -158,11 +151,7 @@ void Esp32NcpClient::destroy() {
     }
     parser_.destroy();
     muxerAtStream_.reset();
-#if PLATFORM_ID == PLATFORM_ARGON || PLATFORM_ID == PLATFORM_ASOM
     serial_.reset();
-#elif PLATFORM_ID == PLATFORM_TRACKER
-    sdio_.reset();
-#endif
 }
 
 int Esp32NcpClient::on() {
@@ -204,11 +193,7 @@ int Esp32NcpClient::enable() {
     if (ncpState_ != NcpState::DISABLED) {
         return 0;
     }
-#if PLATFORM_ID == PLATFORM_ARGON || PLATFORM_ID == PLATFORM_ASOM
     serial_->enabled(true);
-#elif PLATFORM_ID == PLATFORM_TRACKER
-    sdio_->enabled(true);
-#endif
     muxerAtStream_->enabled(true);
     ncpState_ = prevNcpState_;
     off();
@@ -224,11 +209,7 @@ void Esp32NcpClient::disable() {
     }
     prevNcpState_ = state;
     ncpState_ = NcpState::DISABLED;
-#if PLATFORM_ID == PLATFORM_ARGON || PLATFORM_ID == PLATFORM_ASOM
     serial_->enabled(false);
-#elif PLATFORM_ID == PLATFORM_TRACKER
-    sdio_->enabled(false);
-#endif
     muxerAtStream_->enabled(false);
 }
 
@@ -491,16 +472,15 @@ int Esp32NcpClient::waitReady() {
     }
     muxer_.stop();
     espReset();
-#if PLATFORM_ID == PLATFORM_ARGON || PLATFORM_ID == PLATFORM_ASOM
+#if !HAL_PLATFORM_WIFI_NCP_SDIO
     CHECK(serial_->setBaudRate(ESP32_NCP_DEFAULT_SERIAL_BAUDRATE));
+#else
+    // SDIO bus initialization only performs right after power on
+    CHECK(serial_->init());
+#endif // !HAL_PLATFORM_WIFI_NCP_SDIO
+
     CHECK(initParser(serial_.get()));
     skipAll(serial_.get(), 1000);
-#elif PLATFORM_ID == PLATFORM_TRACKER
-    // SDIO bus initialization only performs right after power on
-    CHECK(sdio_->init());
-    CHECK(initParser(sdio_.get()));
-    skipAll(sdio_.get(), 1000);
-#endif
 
     parser_.reset();
     const unsigned timeout = 10000;
@@ -521,11 +501,7 @@ int Esp32NcpClient::waitReady() {
     }
 
     if (ready_) {
-#if PLATFORM_ID == PLATFORM_ARGON || PLATFORM_ID == PLATFORM_ASOM
         skipAll(serial_.get(), 1000);
-#elif PLATFORM_ID == PLATFORM_TRACKER
-        skipAll(sdio_.get(), 1000);
-#endif
         parser_.reset();
         parserError_ = 0;
         LOG(TRACE, "NCP ready to accept AT commands");
@@ -578,11 +554,7 @@ int Esp32NcpClient::initReady() {
 
 int Esp32NcpClient::initMuxer() {
     // Initialize muxer
-#if PLATFORM_ID == PLATFORM_ARGON || PLATFORM_ID == PLATFORM_ASOM
     muxer_.setStream(serial_.get());
-#elif PLATFORM_ID == PLATFORM_TRACKER
-    muxer_.setStream(sdio_.get());
-#endif
     muxer_.setMaxFrameSize(ESP32_NCP_MAX_MUXER_FRAME_SIZE);
     muxer_.setKeepAlivePeriod(ESP32_NCP_KEEPALIVE_PERIOD * 2);
     muxer_.setKeepAliveMaxMissed(ESP32_NCP_KEEPALIVE_MAX_MISSED);
