+++ hal/src/argon/hal_platform_config.h
@@ -31,5 +31,6 @@
 #define HAL_PLATFORM_PMIC_BQ24195_FAULT_COUNT_THRESHOLD (4)
 #define HAL_PLATFORM_RADIO_ANTENNA_EXTERNAL (1)
 #define HAL_PLATFORM_FLASH_MX25L3233F (1)
+#define HAL_PLATFORM_WIFI_COMPAT (1)
 
 #endif // PLATFORM_ID != PLATFORM_ARGON
