+++ hal/src/tracker/hal_platform_config.h
@@ -41,4 +41,8 @@
 
 #define HAL_PLATFORM_FLASH_MX25R6435FZNIL0 (1)
 
+#define HAL_PLATFORM_NCP_COUNT (2)
+#define HAL_PLATFORM_WIFI (1)
+#define HAL_PLATFORM_WIFI_NCP_SDIO (1)
+
 #define PRODUCT_SERIES "boron"
