+++ hal/src/tracker/include.mk
@@ -9,5 +9,9 @@ ifneq (,$(findstring platform,$(DEPENDENCIES)))
 INCLUDE_DIRS += $(HAL_MODULE_PATH)/network/lwip/cellular
 endif
 
+ifneq (,$(findstring platform,$(DEPENDENCIES)))
+INCLUDE_DIRS += $(HAL_MODULE_PATH)/network/lwip/esp32
+INCLUDE_DIRS += $(HAL_MODULE_PATH)/network/ncp/wifi
+endif
 
 include $(TARGET_HAL_PATH)/src/nRF52840/include.mk
