+++ hal/src/tracker/platform_ncp_quectel.cpp
@@ -20,6 +20,7 @@
 #include "exflash_hal.h"
 
 #include "dct.h"
+#include "check.h"
 
 namespace {
 
@@ -51,4 +52,17 @@ PlatformNCPIdentifier platform_primary_ncp_identifier() {
         }
     }
     return (PlatformNCPIdentifier)ncpId;
+}
+
+int platform_ncp_get_info(int idx, PlatformNCPInfo* info) {
+    CHECK_TRUE(info, SYSTEM_ERROR_INVALID_ARGUMENT);
+    CHECK_TRUE(idx > 0 && idx < platform_ncp_count(), SYSTEM_ERROR_INVALID_ARGUMENT);
+    if (idx == 0) {
+        info->identifier = platform_primary_ncp_identifier();
+        info->updatable = false;
+    } else if (idx == 1) {
+        info->identifier = PLATFORM_NCP_ESP32;
+        info->updatable = true;
+    }
+    return 0;
 }
