+++ hal/src/tracker/sources.mk
@@ -14,5 +14,14 @@ CPPSRC += $(call here_files,$(HAL_MODULE_PATH)/network/ncp/cellular/,*.cpp)
 CSRC += $(call here_files,$(HAL_MODULE_PATH)/network/ncp_client/quectel/,*.c)
 CPPSRC += $(call here_files,$(HAL_MODULE_PATH)/network/ncp_client/quectel/,*.cpp)
 
+CSRC += $(call here_files,$(HAL_MODULE_PATH)/network/lwip/esp32/,*.c)
+CPPSRC += $(call here_files,$(HAL_MODULE_PATH)/network/lwip/esp32/,*.cpp)
+
+CSRC += $(call here_files,$(HAL_MODULE_PATH)/network/ncp/wifi/,*.c)
+CPPSRC += $(call here_files,$(HAL_MODULE_PATH)/network/ncp/wifi/,*.cpp)
+
+CSRC += $(call here_files,$(HAL_MODULE_PATH)/network/ncp_client/esp32/,*.c)
+CPPSRC += $(call here_files,$(HAL_MODULE_PATH)/network/ncp_client/esp32/,*.cpp)
+
 include $(HAL_PLATFORM_SRC_PATH)/../nRF52840/sources.mk
