+++ modules/tracker/system-part1/module_system_part1_export.ld
@@ -39,3 +39,4 @@ PROVIDE ( dynalib_location_hal_ble = system_part1_module_table + 96 );
 PROVIDE ( dynalib_location_hal_nfc = system_part1_module_table + 100 );
 PROVIDE ( dynalib_location_hal_mesh_deprecated = system_part1_module_table + 104 );
 PROVIDE ( dynalib_location_hal_posix_syscall = system_part1_module_table + 108 );
+PROVIDE ( dynalib_location_hal_wlan = system_part1_module_table + 112 );
